backend/requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0

# AI - Groq
groq>=0.4.0

# Document Processing
PyPDF2>=3.0.0
pdfplumber>=0.11.0
python-docx>=1.1.0
Pillow>=10.0.0
pytesseract>=0.3.10
pdf2image>=1.16.0

# Data Processing
pandas>=2.0.0

# Logging
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
python-dotenv>=1.0.0

# Google APIs
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.110.0
openpyxl>=3.1.0
